# feat: faster IndexedCrate::new with rayon (#418)